BiasService/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai>=1.21
